Decrease redundant code
The names of images that we test have been extracted into a standalone
function. The version numbers we're testing have been placed into arrays
so we can test multiple versions of the same image in the future. An
image only gets added to the matrix if its Dockerfile has changed and
its name and version are part of the reference list.